requirements.txt
azure-storage-blob
ultralytics
requests
opencv-python
python-dotenv
